# business/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Cargo:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      isSystem:
        type: boolean
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      tenantId:
        type: string
      updatedAt:
        type: string
    type: object
  models.Client:
    properties:
      addresses:
        items:
          $ref: '#/definitions/models.ClientAddress'
        type: array
      contacts:
        description: Relations
        items:
          $ref: '#/definitions/models.Contact'
        type: array
      createdAt:
        type: string
      document:
        description: |-
          Document is a TRANSIENT (gorm:"-", not persisted) field the controller
          populates in API responses after decrypting DocumentEnc. It exists so
          JSON consumers keep seeing a plain "document" field — do not mistake it
          for a duplicate/dead column.
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      phone:
        type: string
      socialName:
        description: |-
          SocialName is exclusive to PF (LGPD — Decreto 8.727/2016). When set, it
          replaces the civil name on every display surface (ticket list, chat
          bubble, header, notifications, Pipeline/Deal, Protocol, reports).
        type: string
      tenantId:
        type: string
      type:
        description: |-
          Type is "pf" (pessoa física) or "pj" (pessoa jurídica). Enum validation
          happens in the controller, not here.
        type: string
      updatedAt:
        type: string
    type: object
  models.ClientAddress:
    properties:
      city:
        type: string
      clientId:
        type: integer
      complement:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      isPrimary:
        type: boolean
      label:
        type: string
      latitude:
        description: |-
          Latitude/Longitude are nullable — populated best-effort by geocoding;
          a failed/unavailable geocode leaves them NULL and never blocks the save.
        type: number
      longitude:
        type: number
      neighborhood:
        type: string
      number:
        type: string
      state:
        type: string
      street:
        type: string
      tenantId:
        type: string
      updatedAt:
        type: string
      zipCode:
        type: string
    type: object
  models.Contact:
    properties:
      client:
        $ref: '#/definitions/models.Client'
      clientId:
        description: |-
          ClientID is nullable — a Contact can exist and generate Tickets without
          ever being linked to a Client (someone messages in without being
          registered). A Contact belongs to at most one Client; the link is
          always manual (ADR 0023) — no unique constraint here, since multiple
          Contacts may point to the same Client.
        type: integer
      createdAt:
        type: string
      email:
        type: string
      groupParticipantCount:
        description: |-
          GroupParticipantCount é preenchido só para IsGroup=true, via
          enriquecimento oportunista (plugins/groups_watch.go
          enrichContactFromGroup) toda vez que o plugin Grupos busca
          GroupInfo.Participants do WhatsApp -- nunca setado para contato
          individual. nil = nunca enriquecido (grupo listado só via mensagem
          recebida, sem passar pelo plugin Grupos ainda).
        type: integer
      id:
        type: integer
      isGroup:
        type: boolean
      lid:
        type: string
      name:
        type: string
      number:
        description: |-
          number e lid são únicos POR TENANT (idx_contacts_tenant_number e
          idx_contacts_tenant_lid, criados em database.addCustomIndexes). O unique
          global antigo (uni_Contacts_number/uni_Contacts_lid) impedia dois tenants
          de terem o mesmo contato e transformava a duplicata em erro 500.
        type: string
      profilePicUrl:
        type: string
      tenantId:
        type: string
      tickets:
        description: Relations
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      updatedAt:
        type: string
      wallet:
        $ref: '#/definitions/models.User'
      walletUserId:
        type: integer
    type: object
...
  models.Message:
    properties:
      ack:
        type: integer
      body:
        type: string
      contactId:
        type: integer
      createdAt:
        type: string
      dataJson:
        type: string
      fromMe:
        type: boolean
      id:
        type: string
      isDeleted:
        type: boolean
      mediaType:
        type: string
      mediaUrl:
        type: string
      participant:
        type: string
      quotedMsg:
        allOf:
        - $ref: '#/definitions/models.Message'
        description: |-
          QuotedMsg é a mensagem citada, preenchida só na listagem (GET /messages/:ticketId) para o
          balão mostrar "em resposta a…". Não é coluna.
      quotedMsgId:
        type: string
      reactions:
        type: string
      read:
        type: boolean
      tenantId:
        type: string
      ticket:
        allOf:
        - $ref: '#/definitions/models.Ticket'
        description: Relations
      ticketId:
        type: integer
      transcription:
        description: |-
          Transcription is filled on-demand when an operator clicks "Transcrever
          áudio" on an audio message (message_transcription.go) — never written
          automatically on receipt, and never sent back to WhatsApp.
        type: string
      updatedAt:
        type: string
    type: object
  models.Permission:
    properties:
      action:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      isSystem:
        type: boolean
      resource:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.Queue:
    properties:
      color:
        description: |-
          Color é cosmético — NÃO é único (o unique global antigo uni_Queues_color
          impedia dois tenants de usarem a mesma cor). Dropado na migração.
        type: string
      createdAt:
        type: string
      distributionStrategy:
        type: string
      greetingMessage:
        type: string
      id:
        type: integer
      name:
        description: |-
          Nome é único POR TENANT (não global) — multi-tenant: dois tenants podem ter
          uma fila "Atendimento Inicial". O unique global antigo (uni_Queues_name)
          quebrava o provisionamento do 2º tenant; é dropado na migração.
        type: string
      parent:
        allOf:
        - $ref: '#/definitions/models.Queue'
        description: Relations
      parentId:
        type: integer
      prioritizeWallet:
        type: boolean
      tenantId:
        type: string
      updatedAt:
        type: string
      whatsapps:
        items:
          $ref: '#/definitions/models.Whatsapp'
        type: array
    type: object
  models.Tag:
    properties:
      archived:
        type: boolean
      color:
        type: string
      createdAt:
        type: string
      description:
        type: string
      group:
        $ref: '#/definitions/models.TagGroup'
      groupId:
        type: integer
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      tenantId:
        type: string
      updatedAt:
        type: string
    type: object
  models.TagGroup:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      tenantId:
        type: string
      updatedAt:
        type: string
    type: object
  models.Tenant:
    properties:
      createdAt:
        type: string
      document:
        type: string
      id:
        type: string
      name:
        type: string
      ownerId:
        type: integer
      status:
        type: string
      updatedAt:
        type: string
      users:
        description: Relations
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.Ticket:
    properties:
      contact:
        allOf:
        - $ref: '#/definitions/models.Contact'
        description: Relations
      contactId:
        type: integer
      createdAt:
        type: string
      id:
        type: integer
      isCommunity:
        type: boolean
      isGroup:
        type: boolean
      isSubGroup:
        type: boolean
      lastMessage:
        type: string
      messages:
        items:
          $ref: '#/definitions/models.Message'
        type: array
      queueId:
        type: integer
      status:
        type: string
      tags:
        description: |-
          Tags is populated manually (batch query via EntityTags) by ListTickets —
          not a GORM association, since EntityTag is polymorphic (entityType+entityId).
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      tenantId:
        type: string
      unreadMessages:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/models.User'
      userId:
        type: integer
      whatsappId:
        type: integer
    type: object
  models.User:
    properties:
      alcance:
        type: string
      cargo:
        $ref: '#/definitions/models.Cargo'
      cargoId:
        type: integer
      configs:
        type: string
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      queues:
        items:
          $ref: '#/definitions/models.Queue'
        type: array
      tenant:
        allOf:
        - $ref: '#/definitions/models.Tenant'
        description: Relations
      tenantId:
        type: string
      tokenVersion:
        type: integer
      updatedAt:
        type: string
      whatsappId:
        type: integer
    type: object
  models.Whatsapp:
    properties:
      battery:
        type: string
      connectionGroupId:
        type: integer
      createdAt:
        type: string
      engineType:
        type: string
      farewellMessage:
        type: string
      firstConnection:
        type: string
      greetingMessage:
        type: string
      id:
        type: integer
      isDefault:
        type: boolean
      izapiaSessionId:
        description: |-
          IzapiaSessionID is the session id assigned by the izapia API (distinct from
          this row's own ID). Only set when EngineType == "izapia".
        type: string
      keepAlive:
        type: boolean
      lastConnectedAt:
        type: string
      lastRiskAction:
        type: string
      lastRiskAt:
        type: string
      lastRiskCode:
        description: |-
          LastRisk* persist the most recent whatsmeow ban/throttle risk signal
          (IQ codes 401/403/429/463) for this connection, so the Conexões page can
          warn the user even after the live SSE event has been missed.
        type: integer
      lastRiskMessage:
        type: string
      name:
        type: string
      number:
        type: string
      plugged:
        type: boolean
      profilePicUrl:
        type: string
      proxyGroupId:
        type: integer
      proxyId:
        type: integer
      proxyMode:
        description: |-
          Proxy assignment (anti-ban). ProxyMode: "none" | "single" | "group".
            single → ProxyID fixo. group → ProxyGroupID + ProxyID guarda o pick atual
            (sticky) ou o último usado (rotate).
        type: string
      qrcode:
        type: string
      queues:
        items:
          $ref: '#/definitions/models.Queue'
        type: array
      retries:
        type: integer
      session:
        type: string
      status:
        type: string
      syncHistory:
        type: boolean
      syncPeriod:
        type: string
      tenantId:
        type: string
      tickets:
        description: Relations
        items:
          $ref: '#/definitions/models.Ticket'
        type: array
      updatedAt:
        type: string
      wid:
        type: string
    type: object
...
host: localhost:8082
info:
  contact:
    email: dev@watink.com.br
    name: Watink Dev
  description: API do backend Go — atendimento e automação WhatsApp.
  license:
    name: Proprietary
    url: https://watink.com.br
  termsOfService: https://watink.com.br/terms
  title: Watink Business API
  version: 2.0.0
paths:
...
  /clients/{id}/contacts/{contactId}:
    delete:
      parameters:
      - description: ID do cliente
        in: path
        name: id
        required: true
        type: integer
      - description: ID do contato
        in: path
        name: contactId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Desvincular contato do cliente
      tags:
      - clients
    post:
      consumes:
      - application/json
      parameters:
      - description: ID do cliente
        in: path
        name: id
        required: true
        type: integer
      - description: ID do contato
        in: path
        name: contactId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Contact'
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Vincular contato ao cliente
      tags:
      - clients
...
  /contacts:
    get:
      parameters:
      - description: Busca por nome ou número
        in: query
        name: searchParam
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Listar contatos
      tags:
      - contacts
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do contato
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar contato
      tags:
      - contacts
  /contacts/{contactId}:
    delete:
      parameters:
      - description: ID do contato
        in: path
        name: contactId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remover contato
      tags:
      - contacts
    get:
      parameters:
      - description: ID do contato
        in: path
        name: contactId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Detalhar contato
      tags:
      - contacts
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do contato
        in: path
        name: contactId
        required: true
        type: integer
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Atualizar contato
      tags:
      - contacts
...
  /contacts/all:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Excluir todos os contatos do tenant
      tags:
      - contacts
  /contacts/bulk-delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: Lista de IDs a excluir
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Excluir contatos selecionados
      tags:
      - contacts
...
  /messages/{ticketId}:
    get:
      parameters:
      - description: ID do ticket
        in: path
        name: ticketId
        required: true
        type: integer
      - description: Página
        in: query
        name: pageNumber
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Listar mensagens do ticket
      tags:
      - messages
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Envia mensagem de texto ou mídia via engine WhatsApp
      parameters:
      - description: ID do ticket
        in: path
        name: ticketId
        required: true
        type: integer
      - description: Corpo JSON para texto
        in: body
        name: body
        schema:
          type: object
      - description: Arquivo de mídia (multipart)
        in: formData
        name: medias
        type: file
      - description: true/false
        in: formData
        name: fromMe
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Enviar mensagem
      tags:
      - messages
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer JWT token. Formato: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"